check validate, merge code, get part policy

+++ Project/demo/src/main/java/com/fptu/swp391/se1839/oemevwarrantymanagement/dto/response/PartInventoryResponse.java
@@ -21,6 +21,6 @@ public class PartInventoryResponse {
     Long serviceCenterId;
     String serviceCenterName;
     String serviceCenterAddress;
-    int quantity;
+    long quantity;
     String unit;
 }
